Move and improve devlinks.rs
Since devlinks will be used by dbus_api in the next commit, move it from
engine/strat_engine/ to engine/.

Add filesystem_mount_path() and pool_directory() fns (the former is
also pub since dbus_api will be using it. (Mulhern)

Implement AsRef for Name to make calling devlinks fns nicer. (Mulhern)

Move DEV_PATH up and out of strat_engine/.
