+++ src/engine/devlinks.rs
@@ -11,7 +11,7 @@ use std::{fs, str};
 use stratis::StratisResult;
 
 use super::super::engine::Pool;
-use super::super::types::{Name, PoolUuid};
+use super::types::{Name, PoolUuid};
 
 use super::engine::DEV_PATH;
 
@@ -44,7 +44,7 @@ pub fn setup_devlinks<'a, I: Iterator<Item = &'a (Name, PoolUuid, &'a Pool)>>(
             pool_added(pool_name)?;
         }
 
-        let pool_path: PathBuf = vec![DEV_PATH, pool_name].iter().collect();
+        let pool_path = pool_directory(pool_name);
 
         let mut existing_files = fs::read_dir(pool_path)?
             .map(|dir_e| {
@@ -71,22 +71,22 @@ pub fn setup_devlinks<'a, I: Iterator<Item = &'a (Name, PoolUuid, &'a Pool)>>(
 
 /// Create a directory when a pool is added.
 pub fn pool_added(pool: &str) -> StratisResult<()> {
-    let p: PathBuf = vec![DEV_PATH, pool].iter().collect();
+    let p = pool_directory(pool);
     fs::create_dir(&p)?;
     Ok(())
 }
 
 /// Remove the directory and its contents when the pool is removed.
 pub fn pool_removed(pool: &str) -> StratisResult<()> {
-    let p: PathBuf = vec![DEV_PATH, pool].iter().collect();
+    let p = pool_directory(pool);
     fs::remove_dir_all(&p)?;
     Ok(())
 }
 
 /// Rename the directory to match the pool's new name.
 pub fn pool_renamed(old_name: &str, new_name: &str) -> StratisResult<()> {
-    let old: PathBuf = vec![DEV_PATH, old_name].iter().collect();
-    let new: PathBuf = vec![DEV_PATH, new_name].iter().collect();
+    let old = pool_directory(old_name);
+    let new = pool_directory(new_name);
     fs::rename(&old, &new)?;
     Ok(())
 }
@@ -94,7 +94,7 @@ pub fn pool_renamed(old_name: &str, new_name: &str) -> StratisResult<()> {
 /// Create a symlink to the new filesystem's block device within its pool's
 /// directory.
 pub fn filesystem_added(pool_name: &str, fs_name: &str, devnode: &Path) -> StratisResult<()> {
-    let p: PathBuf = vec![DEV_PATH, pool_name, fs_name].iter().collect();
+    let p = filesystem_mount_path(pool_name, fs_name);
 
     // Remove existing and recreate to ensure it points to the correct devnode
     let _ = fs::remove_file(&p);
@@ -104,15 +104,29 @@ pub fn filesystem_added(pool_name: &str, fs_name: &str, devnode: &Path) -> Strat
 
 /// Remove the symlink when the filesystem is destroyed.
 pub fn filesystem_removed(pool_name: &str, fs_name: &str) -> StratisResult<()> {
-    let p: PathBuf = vec![DEV_PATH, pool_name, fs_name].iter().collect();
+    let p = filesystem_mount_path(pool_name, fs_name);
     fs::remove_file(&p)?;
     Ok(())
 }
 
 /// Rename the symlink to track the filesystem's new name.
 pub fn filesystem_renamed(pool_name: &str, old_name: &str, new_name: &str) -> StratisResult<()> {
-    let old: PathBuf = vec![DEV_PATH, pool_name, old_name].iter().collect();
-    let new: PathBuf = vec![DEV_PATH, pool_name, new_name].iter().collect();
+    let old = filesystem_mount_path(pool_name, old_name);
+    let new = filesystem_mount_path(pool_name, new_name);
     fs::rename(&old, &new)?;
     Ok(())
+}
+
+/// Given a pool name, synthesize a pool directory name for storing filesystem
+/// mount paths.
+fn pool_directory<T: AsRef<str>>(pool_name: T) -> PathBuf {
+    vec![DEV_PATH, pool_name.as_ref()].iter().collect()
+}
+
+/// Given a pool name and a filesystem name, return the path it should be
+/// available as a device for mounting.
+pub fn filesystem_mount_path<T: AsRef<str>>(pool_name: T, fs_name: T) -> PathBuf {
+    vec![DEV_PATH, pool_name.as_ref(), fs_name.as_ref()]
+        .iter()
+        .collect()
 }
