+++ src/engine/strat_engine/mod.rs
@@ -7,7 +7,6 @@ mod backstore;
 mod cleanup;
 mod cmd;
 mod device;
-mod devlinks;
 mod dm;
 mod dmnames;
 mod engine;
