+++ src/engine/strat_engine/thinpool/mdv.rs
@@ -18,11 +18,11 @@ use devicemapper::{DmDevice, LinearDev, LinearDevTargetParams, TargetLine};
 
 use stratis::StratisResult;
 
+use super::super::super::engine::DEV_PATH;
 use super::super::super::types::{FilesystemUuid, Name, PoolUuid};
 
 use super::super::cmd::create_fs;
 use super::super::dm::get_dm;
-use super::super::engine::DEV_PATH;
 use super::super::serde_structs::FilesystemSave;
 
 use super::filesystem::StratFilesystem;
