+++ src/engine/strat_engine/thinpool/thinpool.rs
@@ -17,6 +17,7 @@ use devicemapper::{
 
 use stratis::{ErrorEnum, StratisError, StratisResult};
 
+use super::super::super::devlinks;
 use super::super::super::engine::Filesystem;
 use super::super::super::structures::Table;
 use super::super::super::types::{FilesystemUuid, Name, PoolUuid, RenameAction};
@@ -24,7 +25,6 @@ use super::super::super::types::{FilesystemUuid, Name, PoolUuid, RenameAction};
 use super::super::backstore::Backstore;
 use super::super::cmd::{thin_check, thin_repair};
 use super::super::device::wipe_sectors;
-use super::super::devlinks;
 use super::super::dm::get_dm;
 use super::super::dmnames::{
     format_flex_ids, format_thin_ids, format_thinpool_ids, FlexRole, ThinPoolRole, ThinRole,
